+++ typed_sql/test/typed_sql/default/int_cast_test/int_cast_test.dart
@@ -12,89 +12,18 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import 'dart:io';
-import 'dart:isolate';
-
-void main() {
-  final outUri = Isolate.resolvePackageUriSync(Uri.parse('package:typed_sql/'))
-      ?.resolve('../test/typed_sql/default/types/');
-  if (outUri == null) {
-    print('Cannot resolve package:typed_sql/');
-    exit(1);
-  }
-
-  for (final i in instances) {
-    final name = i['name'];
-    final target = File.fromUri(
-      outUri.resolve('$name/${name}_test.dart'),
-    );
-    print('Creating: ${target.path}');
-    target.parent.createSync(recursive: true);
-
-    target.writeAsStringSync(i.entries.fold(
-      template,
-      (template, entry) => template.replaceAll('{{${entry.key}}}', entry.value),
-    ));
-  }
-}
-
-final instances = [
-  {
-    'name': 'default_integer',
-    'type': 'int',
-    'defaultValue': '42',
-    'nonDefaultValue': '21',
-  },
-  {
-    'name': 'default_text',
-    'type': 'String',
-    'defaultValue': '\'hello\'',
-    'nonDefaultValue': '\'hello world\'',
-  },
-  {
-    'name': 'default_boolean',
-    'type': 'bool',
-    'defaultValue': 'true',
-    'nonDefaultValue': 'false',
-  },
-  {
-    'name': 'default_real',
-    'type': 'double',
-    'defaultValue': '42.2',
-    'nonDefaultValue': '3.14',
-  },
-  {
-    'name': 'default_zero_real',
-    'type': 'double',
-    'defaultValue': '0.0',
-    'nonDefaultValue': '3.14',
-  },
-  // TODO: Add support for default DateTime
-  /*{
-    'name': 'default_datetime',
-    'type': 'DateTime',
-    'defaultValue': 'DateTime(2024).toUtc()',
-    'nonDefaultValue': 'DateTime(2025).toUtc()',
-  },*/
-];
-
-final template = '''
-// GENERATED CODE - DO NOT MODIFY BY HAND
-//
-// See tool/generate_default_tests.dart
-
 import 'package:typed_sql/typed_sql.dart';
 
-import '../../../testrunner.dart';
+import '../../testrunner.dart';
 
-part '{{name}}_test.g.dart';
+part 'int_cast_test.g.dart';
 
 abstract final class TestDatabase extends Schema {
   Table<Item> get items;
 }
 
-const _defaultValue = {{defaultValue}};
-const _nonDefaultValue = {{nonDefaultValue}};
+const _defaultValue = 42;
+const _nonDefaultValue = 3.14;
 
 @PrimaryKey(['id'])
 abstract final class Item extends Row {
@@ -102,7 +31,7 @@ abstract final class Item extends Row {
   int get id;
 
   @DefaultValue(_defaultValue)
-  {{type}} get value;
+  double get value;
 }
 
 void main() {
@@ -132,7 +61,7 @@ void main() {
         .execute();
 
     final item = await db.items.first.fetch();
-    check(item).isNotNull().value.equals(_defaultValue);
+    check(item).isNotNull().value.equals(_defaultValue.toDouble());
   });
 
   r.addTest('.update() default value', (db) async {
@@ -143,7 +72,7 @@ void main() {
         .execute();
 
     final item = await db.items.first.fetch();
-    check(item).isNotNull().value.equals(_defaultValue);
+    check(item).isNotNull().value.equals(_defaultValue.toDouble());
 
     await db.items
         .byKey(1)
@@ -158,4 +87,3 @@ void main() {
 
   r.run();
 }
-''';