+++ typed_sql/test/typed_sql/default/types/default_text/default_text_test.dart
@@ -4,7 +4,7 @@
 
 import 'package:typed_sql/typed_sql.dart';
 
-import '../../testrunner.dart';
+import '../../../testrunner.dart';
 
 part 'default_text_test.g.dart';